@@ -1,16 +1,15 @@
 name: json_rpc_adapter
 description: HTTP client and shelf server adapters for easier JSON-RPC (optional mirrors for automatic binding).
-version: 0.4.1
+version: 0.5.0
 repository: https://github.com/agilord/json_rpc_adapter
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  http: ^0.12.0
-  json_rpc_2: ^2.0.0
-  meta: ^1.0.0
-  shelf: ^0.7.0
+  http: ^0.13.0
+  json_rpc_2: ^3.0.0
+  shelf: ^1.0.0
 
 dev_dependencies:
   pedantic: ^1.0.0
