@@ -3,15 +3,15 @@ abstract class TestApi {
   Future<String> incrementInt(int x);
 
   Future<void> setter(int x);
-  Future<int> getter();
+  Future<int?> getter();
 }
 
 class TestApiImpl implements TestApi {
-  int _value;
+  int? _value;
 
   @override
   Future<ObjRs> incrementObject(ObjRq rq) async {
-    return ObjRs(rq.value + 1);
+    return ObjRs(rq.value! + 1);
   }
 
   @override
@@ -25,29 +25,29 @@ class TestApiImpl implements TestApi {
   }
 
   @override
-  Future<int> getter() async {
+  Future<int?> getter() async {
     return _value;
   }
 }
 
 class ObjRq {
-  final int value;
+  final int? value;
 
   ObjRq(this.value);
 
   factory ObjRq.fromJson(Map<String, dynamic> json) =>
-      ObjRq(json['value'] as int);
+      ObjRq(json['value'] as int?);
 
   Map<String, dynamic> toJson() => <String, dynamic>{'value': value};
 }
 
 class ObjRs {
-  final int value;
+  final int? value;
 
   ObjRs(this.value);
 
   factory ObjRs.fromJson(Map<String, dynamic> json) =>
-      ObjRs(json['value'] as int);
+      ObjRs(json['value'] as int?);
 
   Map<String, dynamic> toJson() => <String, dynamic>{'value': value};
 }
