@@ -6,7 +6,6 @@ import 'handler.dart';
 
 /// Extends [JsonRpcShelfHandler] with mirror-based API registration.
 extension JsonRpcShelfHandlerMirror on JsonRpcShelfHandler {
-
   /// Registers [T] as an API.
   void registerApi<T>(T instance) {
     final instanz = reflect(instance);
@@ -14,11 +13,12 @@ extension JsonRpcShelfHandlerMirror on JsonRpcShelfHandler {
       registerMethod(method.name, (js) async {
         final param = method.inputIsJsonNative
             ? js
-            : method.inputClass.newInstance(Symbol('fromJson'), [js]).reflectee;
+            : method.inputClass!
+                .newInstance(Symbol('fromJson'), [js]).reflectee;
         final rsf = instanz.invoke(
           Symbol(method.name),
           [if (method.inputClass != null) param],
-        ).reflectee as Future;
+        ).reflectee as Future?;
         final rs = await rsf;
         if (method.outputIsJsonNative) {
           return rs;
